refactor: id 컬럼 추가로 인한 변경 SubscriptionDto 적용 및 리팩터링
