+++ src/test/java/com/sprint/monew/domain/interest/InterestServiceTest.java
@@ -17,8 +17,8 @@ import com.sprint.monew.common.util.CursorPageResponseDto;
 import com.sprint.monew.domain.interest.dto.InterestCreateRequest;
 import com.sprint.monew.domain.interest.dto.InterestDto;
 import com.sprint.monew.domain.interest.dto.InterestUpdateRequest;
-import com.sprint.monew.domain.interest.dto.SubscriptionDto;
 import com.sprint.monew.domain.interest.subscription.Subscription;
+import com.sprint.monew.domain.interest.subscription.SubscriptionDto;
 import com.sprint.monew.domain.interest.subscription.SubscriptionRepository;
 import com.sprint.monew.domain.user.User;
 import com.sprint.monew.domain.user.UserRepository;
